Always track pool-worker PIDs
line_profiler/_child_process_profiling/multiprocessing_patches
    (The `child_pids` optional patch has been reworked into the
    `__pool_worker_pid` mandatory patch.)

    __init__.py::apply()
        Updated docstring
    _infrastructure.py::Registry.from_entry_point()
        Updated doctest
    _mandatory_patches.py
        POOL_WORKER_PID_PATCH
            Migrated from `_optional_patches.py::CHILD_PIDS_PATCH`
        wrap_handle_results(), wrap_process(), wrap_worker()
            Migrated from eponymous functions in `_optional_patches.py`
    _profiling_patches.py
        wrap_worker()
            Now always using the previous `wrap_worker_write_per_task()`
            implementation, because we're no longer catching SIGTERM
            even on POSIX
        wrap_terminate()
            Now only blocking the call to `BaseProcess.terminate()` if
            `self` is a pool worker which has run at least 1 task; this
            avoids having idle workers in a "dirty" state interfering
            with pool termination

line_profiler/rc/line_profiler.toml
::[tool.line_profiler.child_processes.multiprocessing.patches]
    Removed item `child_pids`

setup.py
    Renamed entry point: `line_profiler._multiproc_patches.child_pids`
    -> `.__pool_worker_pid`

tests/test_child_procs.py
    _run_kernprof_main_in_process()
    _test_apply_mp_patches()
    test_apply_mp_patches_success()
        Removed references to the `child_pids` patch
